requirements.txt
django==3.0.8
whitenoise==5.2.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.6
django-db-file-storage==0.5.5

